better handle multiple paths to parse
This changes Parse(string) to Parse(...string) so that it can handle
both a single path, as well as a slice of paths.

This commit also:
- Makes PackageFile exposed externally so that it is a type that
  can be passed around.
- Updates the error message when a package can't be parsed to make
  it more clear that it might not be a valid Python package.
